refactored and centralize segment tracking
