+++ leapp/leapp_graph/traced_node.py
@@ -3,12 +3,13 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
+from __future__ import annotations
+
 from leapp.leapp_graph.leapp_node import LeappNode
 import torch
 import torch.fx as fx
 from torch.fx.proxy import Proxy
-from dataclasses import dataclass, field
-from typing import Any, Literal
+from typing import TYPE_CHECKING
 import operator
 
 from leapp.utils.logging import _get_logger
@@ -35,323 +36,8 @@ try:
 except ImportError:
     wp = None
 
-
-def warp_segment_marker(*inputs):
-    """FX marker for one finalized Warp/APIC segment."""
-    raise RuntimeError(
-        "leapp_warp_segment is an FX marker and should be lowered before execution"
-    )
-
-
-@dataclass
-class WarpTensorRef:
-    # Segment-local canonical name used for APIC params / FX output labels.
-    name: str
-    # Live trace-time object, usually a wp.array, kept for capture/replay work.
-    array: Any
-    # FX proxy that represents this value in the owning TracedTensorNode graph.
-    proxy: Proxy | None = None
-    # Owning LEAPP trace context, normally the TracedTensorNode instance.
-    context: Any | None = None
-    # Detector path showing where the value was found, e.g. args[0] or kwargs['out'].
-    path: str | None = None
-    # Runtime array shape observed during tracing, used for validation/export metadata.
-    shape: tuple | None = None
-    # Runtime dtype observed during tracing, stored as text for lightweight metadata.
-    dtype: str | None = None
-    # Runtime device observed during tracing, stored as text for lightweight metadata.
-    device: str | None = None
-    # Device/host pointer when available; helps dedupe view-like wp.array objects.
-    ptr: int | None = None
-    # True when this ref is a runtime input to the Warp segment.
-    is_input: bool = False
-    # True when this ref is a runtime output of the segment.
-    is_output: bool = False
-    # Index of the detector event that first produced/wrote this ref, when known.
-    produced_by_event_index: int | None = None
-    # Extra detector/export annotations that do not deserve first-class fields yet.
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    @classmethod
-    def from_value(
-        cls,
-        name: str,
-        value: Any,
-        *,
-        path: str | None = None,
-        is_input: bool = False,
-        is_output: bool = False,
-        produced_by_event_index: int | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> "WarpTensorRef":
-        ptr = getattr(value, "ptr", None)
-        try:
-            ptr = int(ptr) if ptr else None
-        except Exception:
-            ptr = None
-
-        shape = getattr(value, "shape", None)
-        if shape is not None:
-            try:
-                shape = tuple(shape)
-            except TypeError:
-                pass
-
-        dtype = getattr(value, "dtype", None)
-        device = getattr(value, "device", None)
-
-        return cls(
-            name=name,
-            array=value,
-            proxy=getattr(value, "proxy", None),
-            context=getattr(value, "context_obj", None),
-            path=path,
-            shape=shape,
-            dtype=str(dtype) if dtype is not None else None,
-            device=str(device) if device is not None else None,
-            ptr=ptr,
-            is_input=is_input,
-            is_output=is_output,
-            produced_by_event_index=produced_by_event_index,
-            metadata=metadata or {},
-        )
-
-    def to_metadata(self) -> dict[str, Any]:
-        return {
-            "name": self.name,
-            "path": self.path,
-            "shape": self.shape,
-            "dtype": self.dtype,
-            "device": self.device,
-            "ptr": self.ptr,
-            "is_input": self.is_input,
-            "is_output": self.is_output,
-            "produced_by_event_index": self.produced_by_event_index,
-            "metadata": self.metadata,
-        }
-
-
-@dataclass
-class WarpSavePlan:
-    """Serialized APIC bundle metadata produced after FX pruning."""
-
-    # Filesystem path to the saved ``.wrp`` file (build-time only).
-    wrp_path: str
-    # Basename of the ``.wrp`` inside the WRPB archive.
-    wrp_name: str
-    input_names: list[str]
-    output_names: list[str]
-    output_shapes: list[list[int]]
-
-
-@dataclass
-class WarpSegment:
-    # Owning LEAPP node name. A segment should not span multiple node graphs.
-    node_name: str
-    # Lifecycle state; invalid segments fail closed instead of silently exporting.
-    status: Literal["open", "closed", "invalid"] = "open"
-    # Detector-recorded top-level Warp events/calls that make up the segment.
-    events: list[Any] = field(default_factory=list)
-    # Runtime APIC/FX inputs, keyed by segment-local canonical name.
-    input_refs: dict[str, WarpTensorRef] = field(default_factory=dict)
-    # Runtime APIC/FX outputs that will get marker-derived proxies.
-    output_refs: dict[str, WarpTensorRef] = field(default_factory=dict)
-    # Conservative possible outputs seen by the detector, not yet confirmed.
-    output_candidates: list[WarpTensorRef] = field(default_factory=list)
-    # FX proxy for the single segment marker node.
-    marker_proxy: Proxy | None = None
-    # Per-output FX proxies derived from marker_proxy.
-    output_proxies: dict[str, Proxy] = field(default_factory=dict)
-    # Extra segment annotations such as capture strategy or detector details.
-    metadata: dict[str, Any] = field(default_factory=dict)
-    # Warp device used for capture/replay when known.
-    device: str | None = None
-    # FX marker name, e.g. ``warp_segment_0``.
-    proxy_name: str | None = None
-    # Saved .wrp path once export writes the APIC bundle.
-    wrp_path: str | None = None
-    # Populated by ``_save_warp_captures`` after pruning.
-    save_plan: WarpSavePlan | None = None
-    # Live APIC graph object during trace/export; intentionally not serialized.
-    apic_graph: Any | None = None
-
-    @property
-    def is_open(self) -> bool:
-        return self.status == "open"
-
-    @property
-    def is_empty(self) -> bool:
-        return not (
-            self.events
-            or self.input_refs
-            or self.output_refs
-            or self.output_candidates
-        )
-
-    def add_event(self, event: Any) -> None:
-        self.events.append(event)
-
-    def add_input_ref(
-        self,
-        value: Any,
-        *,
-        path: str | None = None,
-        produced_by_event_index: int | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> WarpTensorRef:
-        ref = self._coerce_ref(
-            value,
-            name=self._default_ref_name("input", len(self.input_refs)),
-            path=path,
-            is_input=True,
-            produced_by_event_index=produced_by_event_index,
-            metadata=metadata,
-        )
-        existing = self._find_ref(ref, self.input_refs.values())
-        if existing is not None:
-            return self._merge_ref(existing, ref, is_input=True)
-
-        self.input_refs[ref.name] = ref
-        return ref
-
-    def add_output_ref(
-        self,
-        value: Any,
-        *,
-        path: str | None = None,
-        produced_by_event_index: int | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> WarpTensorRef:
-        ref = self._coerce_ref(
-            value,
-            name=self._default_ref_name("output", len(self.output_refs)),
-            path=path,
-            is_output=True,
-            produced_by_event_index=produced_by_event_index,
-            metadata=metadata,
-        )
-        existing = self._find_ref(ref, self.output_refs.values())
-        if existing is not None:
-            return self._merge_ref(existing, ref, is_output=True)
-
-        self.output_refs[ref.name] = ref
-        return ref
-
-    def add_output_candidate(
-        self,
-        value: Any,
-        *,
-        path: str | None = None,
-        produced_by_event_index: int | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> WarpTensorRef:
-        ref = self._coerce_ref(
-            value,
-            name=self._default_ref_name(
-                "output_candidate", len(self.output_candidates)
-            ),
-            path=path,
-            produced_by_event_index=produced_by_event_index,
-            metadata=metadata,
-        )
-        existing = self._find_ref(ref, self.output_candidates)
-        if existing is not None:
-            return self._merge_ref(existing, ref)
-
-        self.output_candidates.append(ref)
-        return ref
-
-    def _coerce_ref(
-        self,
-        value: Any,
-        *,
-        name: str,
-        path: str | None = None,
-        is_input: bool = False,
-        is_output: bool = False,
-        produced_by_event_index: int | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> WarpTensorRef:
-        if isinstance(value, WarpTensorRef):
-            ref = value
-            ref.name = name
-            if path is not None:
-                ref.path = path
-            if metadata:
-                ref.metadata.update(metadata)
-            ref.is_input = ref.is_input or is_input
-            ref.is_output = ref.is_output or is_output
-            if produced_by_event_index is not None:
-                ref.produced_by_event_index = produced_by_event_index
-            return ref
-
-        return WarpTensorRef.from_value(
-            name,
-            value,
-            path=path,
-            is_input=is_input,
-            is_output=is_output,
-            produced_by_event_index=produced_by_event_index,
-            metadata=metadata,
-        )
-
-    @staticmethod
-    def _default_ref_name(prefix: str, index: int) -> str:
-        return f"{prefix}_{index}"
-
-    @staticmethod
-    def _ref_key(ref: WarpTensorRef) -> tuple[int, int | None]:
-        return (id(ref.array), ref.ptr)
-
-    def _find_ref(
-        self, ref: WarpTensorRef, existing_refs: collections.abc.Iterable
-    ) -> WarpTensorRef | None:
-        ref_key = self._ref_key(ref)
-        for existing in existing_refs:
-            if self._ref_key(existing) == ref_key:
-                return existing
-        return None
-
-    @staticmethod
-    def _merge_ref(
-        existing: WarpTensorRef,
-        incoming: WarpTensorRef,
-        *,
-        is_input: bool = False,
-        is_output: bool = False,
-    ) -> WarpTensorRef:
-        existing.is_input = existing.is_input or is_input or incoming.is_input
-        existing.is_output = existing.is_output or is_output or incoming.is_output
-        if existing.path is None:
-            existing.path = incoming.path
-        if existing.proxy is None:
-            existing.proxy = incoming.proxy
-        if existing.context is None:
-            existing.context = incoming.context
-        if existing.produced_by_event_index is None:
-            existing.produced_by_event_index = incoming.produced_by_event_index
-        existing.metadata.update(incoming.metadata)
-        return existing
-
-    def invalidate(self) -> None:
-        self.status = "invalid"
-
-    def to_metadata(self) -> dict[str, Any]:
-        return {
-            "node_name": self.node_name,
-            "status": self.status,
-            "device": self.device,
-            "proxy_name": self.proxy_name,
-            "wrp_path": self.wrp_path,
-            "has_apic_graph": self.apic_graph is not None,
-            "inputs": [ref.to_metadata() for ref in self.input_refs.values()],
-            "outputs": [ref.to_metadata() for ref in self.output_refs.values()],
-            "output_candidates": [
-                ref.to_metadata() for ref in self.output_candidates
-            ],
-            "event_count": len(self.events),
-            "metadata": self.metadata,
-        }
+if TYPE_CHECKING:
+    from leapp.leapp_graph.warp_segment import WarpSegment, WarpTensorRef
 
 
 class TracedTensorNode(LeappNode):
@@ -364,7 +50,7 @@ class TracedTensorNode(LeappNode):
         self.tracer.graph = self.graph
         self.tracer.root = torch.nn.Module()
         self.tracer.tensor_attrs = {}
-        self.warp_segments: list[WarpSegment] = []
+        self._warp_segment_counter = 0
         self._warp_bundle_dirs: list[str] = []
 
         self.m = None
@@ -1013,10 +699,14 @@ class TracedTensorNode(LeappNode):
         if wp is None:
             return
 
-        for segment in self.warp_segments:
+        from leapp.leapp_graph.custom_operator_registry.warp_bundle import (
+            iter_warp_segments_from_graph,
+        )
+
+        for segment in iter_warp_segments_from_graph(self.graph):
             if segment.status != "closed" or segment.apic_graph is None:
                 continue
-            if segment.save_plan is not None:
+            if segment.wrp_path is not None:
                 continue
 
             op_node = (
@@ -1064,13 +754,10 @@ class TracedTensorNode(LeappNode):
             ]
 
             segment.wrp_path = wrp_path
-            segment.save_plan = WarpSavePlan(
-                wrp_path=wrp_path,
-                wrp_name=wrp_name,
-                input_names=input_names,
-                output_names=output_names,
-                output_shapes=output_shapes,
-            )
+            segment.wrp_name = wrp_name
+            segment.input_names = input_names
+            segment.output_names = output_names
+            segment.output_shapes = output_shapes
 
             _get_logger().debug(
                 f"[{self.name}] Saved Warp APIC bundle for '{proxy_name}' at "
@@ -1152,60 +839,8 @@ class TracedTensorNode(LeappNode):
             result[name] = state_info["output"]
         return result
 
-    def add_warp_segment(
-        self,
-        *,
-        device: str | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> WarpSegment:
-        """Create a new open Warp segment for this node.
-
-        If the previous segment is still open, finalize it into an FX marker
-        before starting the next one. This keeps the FX graph at one marker per
-        Warp segment while leaving full model compilation to compile_trace().
-        """
-        if self.warp_segments and self.warp_segments[-1].is_open:
-            self.close_warp_segment(self.warp_segments[-1])
-
-        segment = WarpSegment(
-            node_name=self.name,
-            device=device,
-            metadata=metadata or {},
-        )
-        self.warp_segments.append(segment)
-        return segment
-
-    def get_warp_segment(
-        self,
-        *,
-        device: str | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> WarpSegment:
-        """Return the latest open Warp segment, creating one if needed."""
-        if not self.warp_segments or not self.warp_segments[-1].is_open:
-            return self.add_warp_segment(device=device, metadata=metadata)
-
-        segment = self.warp_segments[-1]
-        if device is not None:
-            segment.device = device
-        if metadata:
-            segment.metadata.update(metadata)
-        return segment
-
-    def close_warp_segment(self, segment: WarpSegment | None = None) -> WarpSegment:
-        """Finalize a Warp segment into a single FX marker.
-
-        The live ``WarpSegment`` remains the source of truth during tracing. The
-        FX nodes keep pointers back to the segment/ref objects, and a single
-        metadata snapshot is written after finalization for debugging/export.
-        """
-        if segment is None:
-            if not self.warp_segments:
-                raise RuntimeError(
-                    f"Node '{self.name}' has no Warp segment to close."
-                )
-            segment = self.warp_segments[-1]
-
+    def insert_warp_marker(self, segment: WarpSegment) -> WarpSegment:
+        """Emit FX nodes for a finalized Warp segment into this node's graph."""
         if segment.status == "closed":
             return segment
         if segment.status == "invalid":
@@ -1216,17 +851,12 @@ class TracedTensorNode(LeappNode):
                 f"'{self.name}'."
             )
 
-        for segment_index, known_segment in enumerate(self.warp_segments):
-            if known_segment is segment:
-                break
-        else:
-            raise ValueError("Warp segment is not registered on this node.")
-
         if segment.is_empty:
             segment.status = "closed"
             return segment
 
-        op_name = f"warp_segment_{segment_index}"
+        op_name = f"warp_segment_{self._warp_segment_counter}"
+        self._warp_segment_counter += 1
         segment.proxy_name = op_name
         input_proxies = [
             ref.proxy for ref in segment.input_refs.values() if ref.proxy is not None
